creates msm.elements package and moves MsmNoteElement there
