+++ history.md
@@ -5,7 +5,7 @@
 - New methods for class `meico.mei.Helper`: adds variant for `findChild()` that finds child elements by attribute value, adds variant for `findSibling()` that finds sibling by ID, variant for `cloneElement()` that creates deep or flat element copies, variant for `createElement()` that creates elements with short UUID-based IDs, `removeAllAttributes()` removes all except specified attributes from element, `appendChildAfterSibling()` inserts child after sibling,  adds variants for `addUUID()` for short UUID-based xml:id attributes with or without namespace, `shiftNoteDiatonicly()` shifts note pitch diatonically, `getHalfstepsFromC()` returns semitones from C, `getHalfstepsBetween()` calculates semitone distance.
 - New class `meico.xml.RichElement` as an Attribute-first base wrapper for XML element access with methods for: attribute access (`get()`, `set()`, `has()`, `is()`, `getAsDouble()`, `getAsInteger()`, `copyValue()`), DOM manipulation (`appendChild()`, `getParent()`, `removeParent()`), child access (`getChildren()`, `getChildrenOfType()`, `getFirstChildByName()`, `getFromChild()`), and ID management (`getId()`, `setId()`, `createNewId()`). Supports `.ges` attribute handling and deep copying.
 - New class `meico.mei.MeiElementHelper` extending RichElement with MEI namespace handling.
-- New class `meico.msm.MsmNoteElement` extending RichElement with MSM-specific methods (`getNoteName()`, `getDuration()`, `isSameNote()`).
+- New class `meico.msm.elements.MsmNoteElement` extending RichElement with MSM-specific methods (`getNoteName()`, `getDuration()`, `isSameNote()`).
 - Adds storing notes in `meico.mpm.elements.maps.data.OrnamentData`.
 - New dictionary `resources/ornaments.dict` for resolving ornament symbols with a list of pitch alterations according to the principal note.
 - New class `meico.mei.ornament.OrnamentExpansion` for managing ornament expansions with methods: `setLabel()`, `addCorrespondence()`, `addElement()`, `append()`, `getNotes()`, `getNote()`, `removeNote()`, `getSegmentLabels()`. Generates `<supplied>` (`getOrnamentExpansionElement()`) elements with `<graceGrp>` (`getGroupElement`) and handles duplicates. Individual segment labels are tracked when multiple ornaments are merged onto the same principal note.
