+++ src/meico/mpm/elements/maps/OrnamentationMap.java
@@ -5,7 +5,7 @@ import meico.mpm.Mpm;
 import meico.mpm.elements.maps.data.OrnamentData;
 import meico.mpm.elements.styles.OrnamentationStyle;
 import meico.mpm.elements.styles.defs.OrnamentDef;
-import meico.msm.MsmNoteElement;
+import meico.msm.elements.MsmNoteElement;
 import meico.xml.RichElement;
 import meico.supplementary.KeyValue;
 import nu.xom.Attribute;
