+++ src/meico/msm/elements/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Contains the classes for MSM elements.
+ * @author Lars Engeln
+ */
+package meico.msm.elements;
